+++ dca-archunit/src/test/java/dev/domaincentric/dca/archunit/fixtures/usecase/good/order/application/shiporder/ShipOrderInputPort.java
@@ -1,0 +1,5 @@
+package dev.domaincentric.dca.archunit.fixtures.usecase.good.order.application.shiporder;
+
+import dev.domaincentric.dca.buildingblocks.hexagonal.port.in.UseCase;
+
+public interface ShipOrderInputPort extends UseCase<ShipOrderCommand, ShipOrderResult> {}